Split the model into separate modules.
- Split autokey.model.__init__.py into separate modules for each class
- Moved class Key and related constants from the iomediator to the model package to break cyclic imports
- Modified both GUIs to work with the new structure
- Fixed failing tests
