+++ lib/autokey/model/modelTypes.py
@@ -1,0 +1,7 @@
+import typing
+
+from .folder import Folder
+from .phrase import Phrase
+from .script import Script
+
+Item = typing.Union[Folder, Phrase, Script]
