+++ lib/autokey/qtui/dialogs/windowfiltersettings.py
@@ -18,11 +18,12 @@ import re
 
 from PyQt5.QtWidgets import QDialog
 
+import autokey.model.folder
+import autokey.model.modelTypes
 from autokey.qtui import common as ui_common
 from .detectdialog import DetectDialog
 
 from autokey import iomediator
-from autokey import model
 
 logger = __import__("autokey.logger").logger.get_logger(__name__)
 
@@ -35,10 +36,10 @@ class WindowFilterSettingsDialog(*ui_common.inherits_from_ui_file_with_name("win
         self.target_item = None
         self.grabber = None  # type: iomediator.WindowGrabber
 
-    def load(self, item: model.Item):
+    def load(self, item: autokey.model.modelTypes.Item):
         self.target_item = item
 
-        if not isinstance(item, model.Folder):
+        if not isinstance(item, autokey.model.folder.Folder):
             self.apply_recursive_check_box.hide()
         else:
             self.apply_recursive_check_box.show()
