+++ lib/autokey/scripting/keyboard.py
@@ -15,6 +15,7 @@
 
 import typing
 
+import autokey.model.phrase
 from autokey import iomediator, model
 
 
@@ -22,12 +23,13 @@ class Keyboard:
     """
     Provides access to the keyboard for event generation.
     """
-    SendMode = model.SendMode
+    SendMode = autokey.model.phrase.SendMode
 
     def __init__(self, mediator):
         self.mediator = mediator  # type: iomediator.IoMediator
 
-    def send_keys(self, key_string, send_mode: typing.Union[model.SendMode, int] = model.SendMode.KEYBOARD):
+    def send_keys(self, key_string, send_mode: typing.Union[
+        autokey.model.phrase.SendMode, int] = autokey.model.phrase.SendMode.KEYBOARD):
         """
         Send a sequence of keys via keyboard events as the default or via clipboard pasting.
         Because the clipboard can only contain
@@ -48,7 +50,7 @@ class Keyboard:
         send_mode = _validate_send_mode(send_mode)
         self.mediator.interface.begin_send()
         try:
-            if send_mode is model.SendMode.KEYBOARD:
+            if send_mode is autokey.model.phrase.SendMode.KEYBOARD:
                 self.mediator.send_string(key_string)
             else:
                 self.mediator.paste_string(key_string, send_mode)
@@ -125,27 +127,25 @@ class Keyboard:
 
 
 def _validate_send_mode(send_mode):
-    permissible_values = "\n".join("keyboard.{}".format(mode) for mode in map(str, model.SendMode))
+    permissible_values = "\n".join("keyboard.{}".format(mode) for mode in map(str, autokey.model.phrase.SendMode))
     if isinstance(send_mode, int):
-        if send_mode in range(len(model.SendMode)):
-            send_mode = tuple(model.SendMode)[send_mode]  # type: model.SendMode
+        if send_mode in range(len(autokey.model.phrase.SendMode)):
+            send_mode = tuple(autokey.model.phrase.SendMode)[send_mode]  # type: model.SendMode
         else:
             permissible_values = "\n".join(
                 "{}: keyboard.{}".format(
-                    number, str(constant)) for number, constant in enumerate(model.SendMode)
+                    number, str(constant)) for number, constant in enumerate(autokey.model.phrase.SendMode)
             )
             raise ValueError(
                 "send_mode out of range for index-based access. "
                 "Permissible values are:\n{}".format(permissible_values))
-    elif isinstance(send_mode, model.SendMode):
-        raise TypeError(
-            "send_mode must be set to an element from keyboard.SendMode (or an interger for index-based access). "
-            "Permissible values are:\n{}".format(permissible_values))
     elif isinstance(send_mode, str):
         try:
-            send_mode = model.SendMode(send_mode)
+            send_mode = autokey.model.phrase.SendMode(send_mode)
         except ValueError as v:
             raise ValueError("Permissible values are: " + permissible_values) from v
-    elif not isinstance(send_mode, model.SendMode):
-        raise TypeError("Unsupported type for send_mode parameter. Use one of: " + permissible_values)
+    elif send_mode is None:  # Selection has value None
+        send_mode = autokey.model.phrase.SendMode.SELECTION
+    elif not isinstance(send_mode, autokey.model.phrase.SendMode):
+        raise TypeError("Unsupported type for send_mode parameter: {} Use one of: {}".format(send_mode, permissible_values))
     return send_mode
