+++ tests/test_iomediator.py
@@ -19,6 +19,7 @@ import pytest
 from hamcrest import *
 
 import autokey.iomediator.constants as iomediator_constants
+import autokey.model.key
 
 
 def generate_tests_for_key_split_re():
@@ -35,6 +36,6 @@ def generate_tests_for_key_split_re():
 @pytest.mark.parametrize("input_string, expected_split", generate_tests_for_key_split_re())
 def test_key_split_re(input_string: str, expected_split: typing.List[str]):
     assert_that(
-        iomediator_constants.KEY_SPLIT_RE.split(input_string),
+        autokey.model.key.KEY_SPLIT_RE.split(input_string),
         contains(*expected_split)
     )
