+++ tests/test_macro.py
@@ -25,27 +25,30 @@ import unittest.mock
 import pytest
 from hamcrest import *
 
+import autokey.model.folder
 import autokey.service
 from autokey.service import PhraseRunner
 from autokey.configmanager.configmanager import ConfigManager
-import autokey.model
 from autokey.scripting import Engine
 
 from autokey.macro import *
 
+
 def get_autokey_dir():
     return os.path.dirname(os.path.realpath(sys.argv[0]))
 
+
 path =  get_autokey_dir() + "/tests/dummy_file.txt"
 
-def create_engine() -> typing.Tuple[Engine, autokey.model.Folder]:
+
+def create_engine() -> typing.Tuple[Engine, autokey.model.folder.Folder]:
     # Make sure to not write to the hard disk
-    test_folder = autokey.model.Folder("Test folder")
+    test_folder = autokey.model.folder.Folder("Test folder")
     test_folder.persist = MagicMock()
 
     # Mock load_global_config to add the test folder to the known folders. This causes the ConfigManager to skip it’s
     # first-run logic.
-    with patch("autokey.model.Phrase.persist"), patch("autokey.model.Folder.persist"),\
+    with patch("autokey.model.phrase.Phrase.persist"), patch("autokey.model.folder.Folder.persist"),\
          patch("autokey.configmanager.configmanager.ConfigManager.load_global_config",
                new=(lambda self: self.folders.append(test_folder))):
         engine = Engine(ConfigManager(MagicMock()), MagicMock(spec=PhraseRunner))
@@ -62,7 +65,7 @@ def create_engine() -> typing.Tuple[Engine, autokey.model.Folder]:
     #     FakeDate.now = classmethod(lambda cls: now(2010, 1, 1))
     #     return date.now()
 class FakeDate(date):
-    "A manipulable date replacement"
+    """A manipulable date replacement"""
     def __new__(cls, *args, **kwargs):
         return date.__new__(date, *args, **kwargs)
 
@@ -89,6 +92,7 @@ def test_arg_parse(test_input, expected, error_msg):
     macro = ScriptMacro(engine)
     assert_that(macro._get_args(test_input), is_(equal_to(expected)),
                 error_msg)
+
 
 # Using date for this because it's the easiest macro with args to expand,
 # without throwing errors on bad args.
@@ -135,12 +139,13 @@ def test_phrase_with_gt_lt_symbols_and_macro(test, expected, error_msg):
     assert_that(expandMacro(engine, test), is_(equal_to(expected)),
                 error_msg)
 
+
 @pytest.mark.skip(reason="For this to work, engine needs to be initialised with a PhraseRunner that isn't a mock. Sadly, that requires an app that isn't a mock.")
 def test_script_macro():
     # Makes use of running script from absolute path.
     engine, folder = create_engine()
-    with patch("autokey.model.Phrase.persist"), patch("autokey.model.Folder.persist"):
-        dummy_folder = autokey.model.Folder("dummy")
+    with patch("autokey.model.phrase.Phrase.persist"), patch("autokey.model.folder.Folder.persist"):
+        dummy_folder = autokey.model.folder.Folder("dummy")
         # This is a duplicate of the phrase added by the target script, but in a
         # different folder.
         dummy = engine.create_phrase(dummy_folder, "arg 1", "arg2", temporary=True)
@@ -152,8 +157,10 @@ def test_script_macro():
         expandMacro(engine, test)
         assert_that(folder.items, has_item(dummy))
 
+
 def test_script_macro_spaced_quoted_args():
     pass
+
 
 def test_cursor_macro():
     engine, folder = create_engine()
@@ -161,6 +168,7 @@ def test_cursor_macro():
     expected="onetwo<left><left><left>"
     assert_that(expandMacro(engine, test), is_(equal_to(expected)),
                 "cursor macro returns wrong text")
+
 
 @pytest.mark.parametrize("test_input, expected, error_msg", [
     ("<cursor><file name={}> types".format(path),
@@ -193,6 +201,7 @@ def test_date_macro():
     assert_that(expandMacro(engine, test), is_(equal_to(expected)),
                 "Date macro fails to expand")
 
+
 def test_file_macro():
     engine, folder = create_engine()
     path =  get_autokey_dir() + "/tests/dummy_file.txt"
@@ -217,8 +226,8 @@ def test_file_macro():
 ])
 def test_macro_expansion(test_input, expected, error_msg):
     engine, folder = create_engine()
-    assert_that(expandMacro(engine, test_input), is_(equal_to(expected)),
-            error_msg)
+    assert_that(expandMacro(engine, test_input), is_(equal_to(expected)), error_msg)
+
 
 def test_system_macro():
     engine, folder = create_engine()
