+++ tests/test_phrase.py
@@ -19,10 +19,11 @@ from unittest.mock import MagicMock
 import pytest
 from hamcrest import *
 
-import autokey.model as model
+import autokey.model.helpers
+import autokey.model.phrase
 from autokey.interface import WindowInfo
 
-ABBR_ONLY = [model.TriggerMode.ABBREVIATION]
+ABBR_ONLY = [autokey.model.helpers.TriggerMode.ABBREVIATION]
 
 AbbreviationType = typing.Union[str, typing.List[str]]
 
@@ -33,7 +34,7 @@ PhraseData = typing.NamedTuple("PhraseData", [
     ("name", str),
     ("abbreviation", AbbreviationType),
     ("content", str),
-    ("trigger_modes", typing.List[model.TriggerMode]),
+    ("trigger_modes", typing.List[autokey.model.helpers.TriggerMode]),
     ("ignore_case", bool),
     ("match_case", bool),
     ("trigger_immediately", bool)
@@ -48,17 +49,17 @@ PhraseResult = typing.NamedTuple("PhraseResult", [
 
 
 def create_phrase(
-        name: str="phrase",
-        abbreviation: AbbreviationType="xp@",
+        name: str = "phrase",
+        abbreviation: AbbreviationType = "xp@",
         content: str = "expansion@autokey.com",  # Values taken from original test code
-        trigger_modes: typing.List[model.TriggerMode]=None,
-        ignore_case: bool=False,
-        match_case: bool=False,
-        trigger_immediately: bool=False) -> model.Phrase:
+        trigger_modes: typing.List[autokey.model.helpers.TriggerMode] = None,
+        ignore_case: bool = False,
+        match_case: bool = False,
+        trigger_immediately: bool = False) -> autokey.model.phrase.Phrase:
     """Save typing by wrapping the Phrase constructor, attribute setters and attributes into a single call."""
     if trigger_modes is None:
-        trigger_modes = [model.TriggerMode.ABBREVIATION]
-    phrase = model.Phrase(name, content)
+        trigger_modes = [autokey.model.helpers.TriggerMode.ABBREVIATION]
+    phrase = autokey.model.phrase.Phrase(name, content)
     if isinstance(abbreviation, str):
         phrase.add_abbreviation(abbreviation)
     else:
@@ -80,7 +81,7 @@ def generate_test_cases_for_ignore_case():
         """Local helper function to save typing constant data"""
         return PhraseData(
             name="name", abbreviation=abbreviation, content=phrase_content,
-            trigger_modes=[model.TriggerMode.ABBREVIATION], ignore_case=ignore_case, match_case=False,
+            trigger_modes=[autokey.model.helpers.TriggerMode.ABBREVIATION], ignore_case=ignore_case, match_case=False,
             trigger_immediately=False)
 
     def phrase_result(expansion_result: str, triggered: bool) -> PhraseResult:
@@ -128,7 +129,7 @@ def generate_test_cases_for_match_case():
         """Local helper function to save typing constant data"""
         return PhraseData(
             name="name", abbreviation=abbreviation, content=phrase_content,
-            trigger_modes=[model.TriggerMode.ABBREVIATION], ignore_case=True, match_case=True,
+            trigger_modes=[autokey.model.helpers.TriggerMode.ABBREVIATION], ignore_case=True, match_case=True,
             trigger_immediately=False)
     
     def phrase_result(expansion_result: str) -> PhraseResult:
@@ -215,7 +216,7 @@ def generate_test_cases_for_trigger_immediately():
         """Local helper function to save typing constant data"""
         return PhraseData(
             name="name", abbreviation=abbreviation, content="Phrase Content.",
-            trigger_modes=[model.TriggerMode.ABBREVIATION], ignore_case=ignore_case, match_case=match_case,
+            trigger_modes=[autokey.model.helpers.TriggerMode.ABBREVIATION], ignore_case=ignore_case, match_case=match_case,
             trigger_immediately=True)
 
     def phrase_result(expansion_result: str, triggered: bool) -> PhraseResult:
@@ -312,7 +313,7 @@ def generate_test_cases_for_case_insensitive_rpartition():
 
 @pytest.mark.parametrize("input_str, match, expected", generate_test_cases_for_case_insensitive_rpartition())
 def test_case_insensitive_rpartition(input_str: str, match: str, expected:typing.Tuple[str, str, str]):
-    assert_that(model.Phrase._case_insensitive_rpartition(input_str, match), is_(equal_to(expected)))
+    assert_that(autokey.model.phrase.Phrase._case_insensitive_rpartition(input_str, match), is_(equal_to(expected)))
 
 
 def generate_test_cases_for_undo_on_backspace():
@@ -322,7 +323,7 @@ def generate_test_cases_for_undo_on_backspace():
         """Local helper function to save typing constant data"""
         return PhraseData(
             name="name", abbreviation="tri", content="ab br",
-            trigger_modes=[model.TriggerMode.ABBREVIATION], ignore_case=False, match_case=False,
+            trigger_modes=[autokey.model.helpers.TriggerMode.ABBREVIATION], ignore_case=False, match_case=False,
             trigger_immediately=trigger_immediately)
 
     def phrase_result(backspace_count: int) -> PhraseResult:
@@ -375,7 +376,7 @@ def generate_test_cases_for_omit_trigger():
         """Local helper function to save typing constant data"""
         return PhraseData(
             name="name", abbreviation="tri", content="ab br",
-            trigger_modes=[model.TriggerMode.ABBREVIATION], ignore_case=False, match_case=False,
+            trigger_modes=[autokey.model.helpers.TriggerMode.ABBREVIATION], ignore_case=False, match_case=False,
             trigger_immediately=trigger_immediately)
 
     def phrase_result(expansion: str, backspace_count: int) -> PhraseResult:
@@ -435,7 +436,7 @@ def generate_test_cases_for_trigger_phrase_inside_word():
         """Local helper function to save typing constant data"""
         return PhraseData(
             name="name", abbreviation="tri", content="ab br",
-            trigger_modes=[model.TriggerMode.ABBREVIATION], ignore_case=False, match_case=False,
+            trigger_modes=[autokey.model.helpers.TriggerMode.ABBREVIATION], ignore_case=False, match_case=False,
             trigger_immediately=trigger_immediately)
 
     def phrase_result(expansion: str, backspace_count: int) -> PhraseResult:
@@ -489,7 +490,7 @@ def generate_test_cases_for_count_lefts_for_cursor_macro():
         """Local helper function to save typing constant data"""
         return PhraseData(
             name="name", abbreviation="tri", content=content,
-            trigger_modes=[model.TriggerMode.ABBREVIATION], ignore_case=False, match_case=False,
+            trigger_modes=[autokey.model.helpers.TriggerMode.ABBREVIATION], ignore_case=False, match_case=False,
             trigger_immediately=trigger_immediately)
 
     def phrase_result(expansion: str, backspace_count: int) -> PhraseResult:
